Javascript: migrate printout functions to their own file
