+++ js/src/pretext-core.js
@@ -18,5 +18,6 @@ import './pretext-dropdown.js';
 import './readability-options.js';
 import '../pretext.js';
 import '../pretext_add_on.js';
+import './pretext-printouts.js';
 import './pretext-embed.js';
 import '../knowl.js';
